requirements.txt
fastapi
uvicorn
paho-mqtt
requests
firebase-admin